feat: add blocks minted by pools query and response handling
- Introduced `GetBlocksMintedByPools` query to retrieve block counts for specified pools.
- Implemented corresponding response structure `BlocksMintedByPools`.
- Updated `EpochActivityCounter` to handle the new query and integrate it with the REST endpoint for pools.
- Modified `State` to include logic for fetching blocks minted based on VRF key hashes.
- Adjusted `PoolExtendedRest` to include blocks minted data in the response.

+++ modules/rest_blockfrost/src/types.rs
@@ -14,7 +14,7 @@ pub struct PoolExtendedRest {
     pub hex: String,
     pub active_stake: String,  // u64 in string
     pub live_stake: String, // u64 in string
-    pub blocks_minted: u32,
+    pub blocks_minted: usize,
     pub live_saturation: f32,
     pub declared_pledge: String, // u64 in string
     pub margin_cost: f32,
